pyproject.toml
[tool.poetry]
name = "librarian"
version = "0.1.0"
description = ""
authors = ["Myles Braithwaite <me@mylesbraithwaite.com>"]
readme = "README.md"
packages = [{include = "librarian"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
datasette = "^0.64.3"
datasette-media = "^0.5.1"
datasette-publish-vercel = "^0.14.2"
datasette-render-image-tags = "^0.1"
datasette-render-markdown = "^2.1.1"
pandas = "^2.0.2"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
pytz = "^2023.3"
requests = "^2.31.0"
sqlite-utils = "^3.30"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
bandit = {extras = ["toml"], version = "^1.7.5"}
black = "^23.1.0"
bpython = "^0.24"
ipython = "^8.11.0"
isort = "^5.12.0"
mypy = "^1.0.1"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
responses = "^0.23.1"
ruff = "^0.0.270"
types-python-dateutil = "^2.8.19.10"
types-pytz = "^2023.3.0.0"
types-requests = "^2.31.0.1"
types-tabulate = "^0.9.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
librarian = "librarian.cli:cli"

[tool.ruff]
line-length = 80

[tool.black]
line-length = 80

[tool.isort]
profile = "black"
line_length = 80

[tool.mypy]

[[tool.mypy.overrides]]
module = ["pandas"]
ignore_missing_imports = true
